+++ scripts/deploy.js
@@ -5,39 +5,52 @@ const path = require('path')
 const chromeDeploy = require('chrome-extension-deploy')
 const firefoxDeploy = require('firefox-extension-deploy')
 
-console.log('Deploying to Chrome Web Store and Add-ons for Firefox...')
+// eslint-disable-next-line no-void
+void (async function() {
+    console.log(
+        '⏳ Deploying to Chrome Web Store and Add-ons for Firefox... ⏳'
+    )
 
-const zipPath = path.join(__dirname, '../extension.zip')
+    const zipPath = path.join(__dirname, '../extension.zip')
 
-chromeDeploy({
-    clientId: process.env.CHROME_CLIENT_ID,
-    clientSecret: process.env.CHROME_CLIENT_SECRET,
-    refreshToken: process.env.CHROME_REFRESH_TOKEN,
-    id: 'afppminkfnfngihdocacbgeajbbdklkf',
-    zip: fs.readFileSync(zipPath),
-}).then(
-    () => {
-        console.log('Chrome deployment complete!')
-    },
-    err => {
-        console.error('Chrome deployment failed: ', err)
+    const chromePromise = chromeDeploy({
+        clientId: process.env.CHROME_CLIENT_ID,
+        clientSecret: process.env.CHROME_CLIENT_SECRET,
+        refreshToken: process.env.CHROME_REFRESH_TOKEN,
+        id: 'afppminkfnfngihdocacbgeajbbdklkf',
+        zip: fs.readFileSync(zipPath),
+    })
+
+    const firefoxPromise = firefoxDeploy({
+        issuer: process.env.FIREFOX_ISSUER,
+        secret: process.env.FIREFOX_SECRET,
+        id: 'refined-bitbucket@refined-bitbucket.org',
+        // eslint-disable-next-line import/no-unresolved
+        version: require('../extension/manifest.json').version, // eslint-disable-line global-require
+        src: fs.createReadStream(zipPath),
+    })
+
+    let chromeSuccess
+    try {
+        await chromePromise
+        console.log('✅ Chrome deployment complete!')
+        chromeSuccess = true
+    } catch (error) {
+        console.error('❌ Chrome deployment failed: ', error)
+        chromeSuccess = false
+    }
+
+    let firefoxSucceess
+    try {
+        await firefoxPromise
+        console.log('✅ Firefox deployment complete!')
+        firefoxSucceess = true
+    } catch (error) {
+        console.error('❌ Firefox failed: ', error)
+        firefoxSucceess = false
+    }
+
+    if (!chromeSuccess || !firefoxSucceess) {
         process.exitCode = 1
     }
-)
-
-firefoxDeploy({
-    issuer: process.env.FIREFOX_ISSUER,
-    secret: process.env.FIREFOX_SECRET,
-    id: 'refined-bitbucket@refined-bitbucket.org',
-    // eslint-disable-next-line import/no-unresolved
-    version: require('../extension/manifest.json').version, // eslint-disable-line global-require
-    src: fs.createReadStream(zipPath),
-}).then(
-    () => {
-        console.log('Firefox deployment complete!')
-    },
-    err => {
-        console.error('Firefox failed: ', err)
-        process.exitCode = 1
-    }
-)
+})()