+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ### Features
 
 -   **New PR UI**. Insert "Copy filename to clipboard" button in diff header,
-    [pull request #xxx](https://github.com/refined-bitbucket/refined-bitbucket/pull/xxx).
+    [pull request #332](https://github.com/refined-bitbucket/refined-bitbucket/pull/332).
 
 # 3.19.0 (2020-01-27)